backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
python-multipart
python-dateutil
pytz
requests
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
firebase-admin
numpy
openai